Refactor UserCreateAPIView
